refactor(deploy): replace runPipeline with Deployer + WalletHandler classes; consolidate seed helpers
Reshapes the deploy entry-point around resource-managed classes:

- `Deployer` (deployer.ts) replaces the procedural `runPipeline`
  function. `Deployer.prepare(opts)` loads config + signing key, starts
  the proof server, builds or adopts a wallet, and loads constructor
  args; the returned instance exposes `.deploy()` and `.dryRun()`. CLI
  flow becomes:

      await using deployer = await Deployer.prepare(opts);
      return dryRun ? await deployer.dryRun() : await deployer.deploy();

  Owned resources are accumulated in an `AsyncDisposableStack` inside
  `prepare`, moved to the instance on success, and disposed in reverse
  order from `[Symbol.asyncDispose]`. Mid-prepare failures unwind
  cleanly via the local `await using` on the stack.

- `WalletHandler` (wallet/handler.ts) replaces the
  `buildDeployerWallet` free function. It wraps a built
  `MidnightWalletProvider` and implements `[Symbol.asyncDispose]`, so
  the Deployer adds it to its stack with a single `stack.use(owned)`
  instead of `stack.defer(...)` with custom try/catch. Mirrors the
  `ProofServer` pattern.

- `wallet/seeds.ts` merges the previous `local-seeds.ts`,
  `normalize.ts`, and `resolve.ts` into one module. None of those owned
  state or a lifecycle, so a class wrapper would be ceremony; a single
  file with three exports is the right unit. Test renamed
  `normalize.test.ts` -> `seeds.test.ts`.

Public API: `runPipeline as deploy` and `PipelineOptions/PipelineResult`
are gone; consumers now use `Deployer`, `DeployerOptions`,
`DeployResult`. `buildDeployerWallet` replaced by `WalletHandler`.
Integration harness and CLI updated.
