+++ tests/integrations/specs/errors.spec.ts
@@ -1,19 +1,19 @@
-import { deploy, ConfigError } from '@openzeppelin/compact-deploy';
+import { ConfigError, Deployer } from '@openzeppelin/compact-deploy';
 import { describe, expect, it } from 'vitest';
 import { testLogger } from '../_harness/logger.ts';
 import { CONFIG_PATH, requireFixtureArtifact } from '../_harness/paths.ts';
 
 /**
- * Spec: pipeline surfaces typed `ConfigError`s for foreseeable user
- * mistakes, with messages that name the offending key/value. These run
- * against the live stack but never get past the config-validation phase,
- * so they're fast.
+ * Spec: Deployer.prepare surfaces typed `ConfigError`s for foreseeable
+ * user mistakes, with messages that name the offending key/value. These
+ * run against the live stack but never get past the config-validation
+ * phase, so they're fast.
  */
 describe('compact-deploy — config errors are typed and actionable', () => {
   it('rejects an unknown contract name', async () => {
     requireFixtureArtifact();
     await expect(
-      deploy({
+      Deployer.prepare({
         contract: 'Nonexistent',
         network: 'local',
         configPath: CONFIG_PATH,
@@ -25,7 +25,7 @@ describe('compact-deploy — config errors are typed and actionable', () => {
   it('rejects an unknown network name', async () => {
     requireFixtureArtifact();
     await expect(
-      deploy({
+      Deployer.prepare({
         contract: 'Counter',
         network: 'unknown-network',
         configPath: CONFIG_PATH,
@@ -36,7 +36,7 @@ describe('compact-deploy — config errors are typed and actionable', () => {
 
   it('rejects a missing compact.toml path', async () => {
     await expect(
-      deploy({
+      Deployer.prepare({
         contract: 'Counter',
         network: 'local',
         configPath: '/nonexistent/compact.toml',
